convert express js files to typescript
