--- server/server.dockerfile
@@ -1,17 +1,0 @@
-FROM node:20
-
-WORKDIR /usr/src/app
-
-COPY package*.json ./
-
-RUN npm install
-
-COPY prisma ./prisma
-
-RUN npx prisma generate
-
-COPY . .
-
-EXPOSE 5000
-
-CMD ["npm", "run", "dev"]
