--- server/swagger-output.json
@@ -1,208 +1,0 @@
-{
-  "swagger": "2.0",
-  "info": {
-    "title": "Chat App API",
-    "description": "Chat App API Documentation",
-    "version": "1.0.0"
-  },
-  "host": "localhost:5000",
-  "basePath": "/api/v1",
-  "tags": [
-    {
-      "name": "Conversations",
-      "description": "Conversation endpoints"
-    }
-  ],
-  "schemes": [
-    "http"
-  ],
-  "consumes": [
-    "application/json"
-  ],
-  "produces": [
-    "application/json"
-  ],
-  "paths": {
-    "/conversation/create": {
-      "post": {
-        
-        "summary": "Create a new conversation",
-        "description":  "Endpoint for creating a new conversation for a user.",
-        "parameters": [
-          {
-            "name": "body",
-            "in": "body",
-            "schema": {
-              "type": "object",
-              "properties": {
-                "userId": {
-                  "example": "6d25380c-3ae8-4023-af50-2dfce1fb8fa4"
-                }
-              }
-            }
-          }
-        ],
-        "responses": {
-          "201": {
-            "description": "Created"
-          },
-          "400": {
-            "description": "Bad Request"
-          },
-          "404": {
-            "description": "Not Found"
-          },
-          "500": {
-            "description": "Internal Server Error"
-          }
-        }
-      }
-    },
-    "/conversation/all/{userId}": {
-      "get": {
-        "summary": "Get all conversations for a user",
-        "description": "Fetch all conversations associated with the given user ID.",
-        "parameters": [
-          {
-            "name": "userId",
-            "in": "path",
-            "required": true,
-            "type": "string"
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "OK"
-          },
-          "400": {
-            "description": "Bad Request"
-          },
-          "404": {
-            "description": "Not Found"
-          },
-          "500": {
-            "description": "Internal Server Error"
-          }
-        }
-      }
-    },
-    "/conversation/{id}": {
-      "delete": {
-        "summary": "Delete a specific conversation",
-        "description": "Delete a conversation by its unique ID.",
-        "parameters": [
-          {
-            "name": "id",
-            "in": "path",
-            "required": true,
-            "type": "string"
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "OK"
-          },
-          "400": {
-            "description": "Bad Request"
-          },
-          "404": {
-            "description": "Not Found"
-          },
-          "500": {
-            "description": "Internal Server Error"
-          }
-        }
-      },
-      "get": {
-        "summary": "Get details of a specific conversation",
-        "description": "Retrieve information about a single conversation using its ID.",
-        "parameters": [
-          {
-            "name": "id",
-            "in": "path",
-            "required": true,
-            "type": "string"
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "OK"
-          },
-          "400": {
-            "description": "Bad Request"
-          },
-          "404": {
-            "description": "Not Found"
-          },
-          "500": {
-            "description": "Internal Server Error"
-          }
-        }
-      }
-    },
-    "/conversation/send": {
-      "post": {
-        "summary": "Send a message to a conversation",
-        "description": "Send a new message to a specified conversation.",
-        "parameters": [
-          {
-            "name": "body",
-            "in": "body",
-            "schema": {
-              "type": "object",
-              "properties": {
-                "content": {
-                  "example": "any"
-                },
-                "conversationId": {
-                  "example": "any"
-                }
-              }
-            }
-          }
-        ],
-        "responses": {
-          "201": {
-            "description": "Created"
-          },
-          "400": {
-            "description": "Bad Request"
-          },
-          "404": {
-            "description": "Not Found"
-          },
-          "500": {
-            "description": "Internal Server Error"
-          }
-        }
-      }
-    }
-  },
-  "definitions": {
-    "Conversation": {
-      "type": "object",
-      "properties": {
-        "id": {
-          "type": "string",
-          "example": "string"
-        },
-        "title": {
-          "type": "string",
-          "example": "string"
-        },
-        "userId": {
-          "type": "string",
-          "example": "string"
-        },
-        "createdAt": {
-          "type": "string",
-          "example": "2023-01-01T00:00:00Z"
-        },
-        "updatedAt": {
-          "type": "string",
-          "example": "2023-01-01T00:00:00Z"
-        }
-      }
-    }
-  }
-}
