+++ server/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "target": "es2016",
+    "module": "commonjs",   
+    "rootDir": "./src",                
+    "outDir": "./dist",              
+    "esModuleInterop": true,            
+    "forceConsistentCasingInFileNames": true,        
+    "strict": true,              
+    "noImplicitAny": true,   
+    "strictNullChecks": true,  
+    "strictFunctionTypes": true,     
+     "skipLibCheck": true    
+  }
+}
