+++ server/tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./src/index.ts"],"version":"5.8.3"}
